fixed numbers in CommonResultSubreddits

+++ Models/SearchResultStatsModel.cs
@@ -4,7 +4,7 @@ namespace GlanceReddit.Models
 {
 	public class SearchResultStatsModel
 	{
-		public Dictionary<string, double> CommonResultSubreddits { get; set; }
+		public Dictionary<string, int> CommonResultSubreddits { get; set; }
 
 		public QueryPopularity SearchPopularity { get; set; }
 	}
